Move brew/restore calc to util file
